Row实现更新由接口改为抽像类
